requirements.txt
Flask==3.0.0
Pillow==8.4.0
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
alembic==1.12.0
flake8==4.0.1
json5==0.9.6
mutagen==1.45.1
pexpect==4.8.0
requests==2.31.0
## The following requirements were added by pip freeze:
aiodataloader==0.4.0
altgraph==0.17.4
aniso8601==9.0.1
attrs==21.4.0
bidict==0.23.1
blinker==1.7.0
boolean.py==4.0
CacheControl==0.13.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.7
cyclonedx-python-lib==4.2.3
Cython==0.29.35
defusedxml==0.7.1
filelock==3.13.1
Flask-SocketIO==5.3.6
gevent==24.2.1
greenlet==3.0.3
h11==0.14.0
html5lib==1.1
idna==3.4
importlib-metadata==6.8.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
license-expression==30.2.0
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mccabe==0.6.1
mdurl==0.1.2
msgpack==1.0.7
packageurl-python==0.13.1
packaging==23.2
pip-api==0.0.30
pip-requirements-parser==32.0.1
pip_audit==2.6.1
pluggy==1.0.0
promise==2.3
ptyprocess==0.7.0
py==1.11.0
py-serializable==0.11.1
pycdio==2.1.1
pycodestyle==2.8.0
pydeps==1.12.17
pyflakes==2.4.0
Pygments==2.17.2
pyparsing==3.0.7
pytest==7.0.1
python-engineio==4.9.0
python-socketio==5.11.1
pytube==12.1.3
rich==13.7.0
Rx==1.6.3
simple-websocket==1.0.0
singledispatch==3.7.0
six==1.16.0
sortedcontainers==2.4.0
stdlib-list==0.9.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.7.0
urllib3==1.26.14
webencodings==0.5.1
Werkzeug==3.0.1
wsproto==1.2.0
youtube.py==2.1.0
zipp==3.17.0
zope.event==5.0
zope.interface==6.2
